backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
groq>=0.15.0
google-search-results>=2.4.2
newspaper3k>=0.2.8
requests>=2.31.0
lxml>=4.9.3
lxml_html_clean>=0.1.0
Pillow>=10.1.0
python-multipart>=0.0.6
